pyproject.toml
[tool.poetry]
name = "os-twitter-bot"
version = "0.1.0"
description = "Twitter bot for Ether Gals on opensea.io"
authors = ["sh0xdh <info@shodh.co>"]
packages = [
  {include = "scripts/"},
  {include = "alembic/"},
  {include = "src/"}
]

[tool.poetry.scripts]
test = "scripts.test:test"
lint = "scripts.lint:lint"
fix = "scripts.lint:fix"
"health-server" = "scripts.server:run"
"twitter-bot" = "scripts.bot:run_bot"
"db:migrate" = "scripts.db:migrate"
"db:rollback" = "scripts.db:rollback"
"db:seed" = "scripts.db:seed"
"db:history" = "scripts.db:migration_history"
"db:generate_migration" = "scripts.db:generate_migration"

[tool.poetry.dependencies]
python = "^3.8"
rich = "^10.10.0"
requests = "^2.26.0"
tweepy = "^3.10.0"
fastcore = "^1.3.26"
click = "^8.0.1"
python-dotenv = "^0.19.0"
pymongo = "^3.12.0"
pymongo-migrate = "^0.11.0"
pydantic = "^1.8.2"
tqdm = "^4.62.2"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-dateutil = "^2.8.2"
ipython = "^7.27.0"
rq-scheduler = "^0.11.0"
alembic = "^1.7.3"
pydantic-sqlalchemy = "^0.0.9"
loguru = "^0.5.3"
schedule = "^1.1.0"

[tool.poetry.dev-dependencies]
vcrpy = "^4.1.1"
ipython = "^7.27.0"
pre-commit = "^2.15.0"
commitizen = "^2.18.1"
autoflake = "^1.4"
isort = "^5.9.3"
black = {version = "^21.9b0", allow-prereleases = true}
pytest = "^6.2.5"
jupyter = "^1.0.0"

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "0.1.0"